update APM implementation
