+++ packages/govern-console/src/App.tsx
@@ -1,14 +1,11 @@
 import Home from 'containers/HomePage/HomePage';
 import { HashRouter as Router } from 'react-router-dom';
-import { WalletProvider } from './providers/AugmentedWallet';
 
 const App = () => {
   return (
-    <WalletProvider>
-      <Router>
-        <Home />
-      </Router>
-    </WalletProvider>
+    <Router>
+      <Home />
+    </Router>
   );
 };
